feat: login, registro, password recovery y voter para post

+++ src/Context/Admin/Post/Repository/PostPersister.php
@@ -3,15 +3,21 @@
 namespace App\Context\Admin\Post\Repository;
 
 use App\Entity\Post;
+use App\Exception\UserDeleteForbiddenException;
+use App\Security\Voter\PostVoter;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationChecker;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 final class PostPersister
 {
     private EntityManagerInterface $em;
+    private AuthorizationCheckerInterface $authorizationChecker;
 
-    public function __construct(EntityManagerInterface $em)
+    public function __construct(EntityManagerInterface $em, AuthorizationCheckerInterface $authorizationChecker)
     {
         $this->em = $em;
+        $this->authorizationChecker = $authorizationChecker;
     }
 
     public function persist(Post $postEntity): void
@@ -23,6 +29,11 @@ final class PostPersister
     public function delete(int $id): void
     {
         $postEntity = $this->em->getReference(Post::class, $id);
+
+        if (!$this->authorizationChecker->isGranted(PostVoter::DELETE, $postEntity)) {
+            throw new UserDeleteForbiddenException('No puedes borrar este post porque no te pertenece');
+        }
+
         $this->em->remove($postEntity);
         $this->em->flush();
     }
